% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lextract}
\alias{lextract}
\title{List extract}
\usage{
lextract(l, pattern, path.only = FALSE, verbose = TRUE, all.paths = FALSE)
}
\arguments{
\item{l}{a list with named element(s)}

\item{pattern}{a pattern to match list element(s)}

\item{path.only}{logical; if \code{TRUE}, matching object is not returned,
only named and unnamed paths}

\item{verbose}{logical; if \code{TRUE}, named and unnamed paths to the
matching elements are printed}

\item{all.paths}{logical; if \code{TRUE}, named and unnamed paths to all
list elements is returned}
}
\description{
Extract a named object from a list without knowing its specific path in
the list. Alternatively, return all list paths.
}
\examples{
fit <- lm(mpg ~ wt, mtcars)
fit$qr$qraux

lextract(fit, 'qraux')
lextract(fit, all.paths = TRUE)

}
